Static ECS

+++ src/ecs/defintions.hpp
@@ -12,10 +12,10 @@
 #include <set>
 #include <stdexcept>
 
-using EntityID = std::size_t;
+using EntityID = std::uint16_t;
 constexpr EntityID MAX_ENTITIES = 10240;
 
-using ComponentID = std::size_t;
+using ComponentID = std::uint8_t;
 constexpr ComponentID MAX_COMPONENTS = 16;
 
 using Signature = std::bitset<MAX_COMPONENTS>;
